backend/requirements.txt
cohere
groq
pypdf
PyMuPDF
langchain
langchain-core
python-dotenv
pinecone
google-search-results
requests
beautifulsoup4
fastapi
uvicorn
transformers
requests
pydantic
python-multipart
passlib[bcrypt]
